Add the `RKDropdownAlert` to the current view controller.
If a view controller was presented as modal, it would obscure the `RKDropdownAlert` from view. Using this method, the alert view is added to the currently visible view controller. Oftentimes, that's still the root view controller (this was tested), but there are some cases where it's a different view controller.